# docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handler.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  model.Brand:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
...
  model.Buyer:
    properties:
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      id:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      password:
        type: string
    type: object
...
  model.Category:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
...
  model.Image:
    properties:
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      item_id:
        type: integer
      url:
        type: string
    type: object
  model.Item:
    properties:
      article:
        type: string
      brand:
        $ref: '#/definitions/model.Brand'
      brand_id:
        type: integer
      category:
        $ref: '#/definitions/model.Category'
      category_id:
        type: integer
      description:
        type: string
      favoritedBy:
        items:
          $ref: '#/definitions/model.Buyer'
        type: array
      height:
        type: integer
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/model.Image'
        type: array
      length:
        type: integer
      material:
        $ref: '#/definitions/model.Material'
      material_id:
        type: integer
      name:
        type: string
      price:
        type: number
      price_with_discount:
        type: number
      quantity:
        type: integer
      seller:
        $ref: '#/definitions/model.Seller'
      seller_id:
        type: string
      weight:
        type: integer
      width:
        type: integer
    type: object
...
  model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.Material:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
...
  model.Order:
    properties:
      buyer:
        $ref: '#/definitions/model.Buyer'
      buyer_id:
        type: string
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      status:
        type: string
      total:
        type: number
    type: object
  model.OrderItem:
    properties:
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      item_id:
        type: integer
      order:
        $ref: '#/definitions/model.Order'
      order_id:
        type: integer
      quantity:
        type: integer
      total:
        type: number
      unit_price:
        type: number
    type: object
  model.Seller:
    properties:
      balance:
        type: number
      email:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
      password:
        type: string
      shop_name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /buyer/order:
    get:
      description: Retrieve an order by its ID for the buyer
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order ID
        in: query
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order data
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get an order by ID
      tags:
      - Orders
    post:
      description: Create an order from the items in the cart of the current buyer
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order created successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Cart is empty
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new order
      tags:
      - Orders
...
  /seller:
    get:
      description: Retrieve seller details by ID
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Seller data
          schema:
            $ref: '#/definitions/model.Seller'
        "403":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to retrieve seller
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get seller information
      tags:
      - Sellers
    patch:
      consumes:
      - application/json
      description: Update the details of the seller
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated seller data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Seller'
      produces:
      - application/json
      responses:
        "200":
          description: Seller updated successfully
          schema:
            type: string
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to update seller
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update seller information
      tags:
      - Sellers
...
  /sign-in/seller:
    post:
      consumes:
      - application/json
      description: Authenticate a seller with their email and password
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Seller data
          schema:
            $ref: '#/definitions/model.Seller'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Seller sign in
      tags:
      - Sellers
...
  /sign-up/seller:
    post:
      consumes:
      - application/json
      description: Register a new seller with the provided details
      parameters:
      - description: Seller data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Seller'
      produces:
      - application/json
      responses:
        "201":
          description: Seller registered successfully
          schema:
            type: string
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to register seller
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register a new seller
      tags:
      - Sellers
swagger: "2.0"